requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
duckdb==1.4.4
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.32
starlette==0.50.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
